requirements.txt
streamlit
gtts
PyPDF2
python-docx